feat(broker): 实现消息持久化和异步刷盘功能- 新增 CommitLogAppendHandler 和 MMapFileMode 类，实现消息的持久化存储 - 添加 EagleMqTopicLoader 和 FileContentReaderUtil 类，支持主题配置的加载和保存 - 实现 ConsumerQueueDetailModel 和 CommitLogFileNameUtil，用于消费队列和日志文件名的管理 - 优化 BrokerApplication 和 CommonCache，支持新的消息持久化功能 - 调整 mq_topic.json 配置文件格式，增加日志文件名和偏移量信息

+++ mq-broker/src/main/java/org/example/broker/model/TopicLatestCommitLogModel.java
@@ -2,6 +2,8 @@ package org.example.broker.model;
 
 import lombok.Data;
 
+import java.util.concurrent.atomic.AtomicLong;
+
 /**
  * @author qushutao
  * @since 2025/5/25-20:45
@@ -13,8 +15,9 @@ public class TopicLatestCommitLogModel {
     private String fileName;
 
     // 文件最大偏移量
-    private int offsetLimit;
+    private long offsetLimit;
 
     // 偏移量最新写入文件的位置
-    private int offset;
+    private AtomicLong offset;
+
 }
